update page penambahan akses 7field

+++ routes/web.php
@@ -80,7 +80,7 @@ Route::middleware('auth','admin')->group(function () {
     Route::put('/admin/transaksi/{user}', [TransaksiController::class, 'update'])->name('admin.transaksi.update');
     Route::delete('/admin/transaksi/{id}', [TransaksiController::class, 'destroy'])->name('admin.transaksi.destroy');
     Route::get('/parkir_masuk', [ParkirMasukController::class, 'index'])->name('parkir.masuk');
-    Route::delete('/parkir_masuk//{id}', [ParkirMasukController::class, 'destroy'])->name('admin.parkir_masuk.destroy');
+    Route::delete('admin/parkir_masuk/{id}', [ParkirMasukController::class, 'destroy'])->name('admin.parkir_masuk.destroy');
    
 });
